refactor(sdk): split user file into modules
